拆分pushboxes
