+++ .gitignore
@@ -1,3 +1,4 @@
+build
 .DS_Store
 *.o
 *.a
@@ -53,3 +54,5 @@ Generated\ Files/
 *.svclog
 *.scc
 *.un~
+*.swp
+*.swp~
