+++ level1/p10_pushBoxes/CMakeLists.txt
@@ -1,0 +1,7 @@
+cmake_minimum_required(VERSION 3.7)
+project(pushboxes)
+aux_source_directory(${PROJECT_SOURCE_DIR}/src DIR_SRC)
+include_directories(${PROJECT_SOURCE_DIR}/include)
+link_directories(${PROJECT_SOURCE_DIR}/lib)
+set(EXECUTABLE_OUTPUT_PATH ${PROJECT_SOURCE_DIR}/bin)
+add_executable(pushboxes ${DIR_SRC})
